+++ NEWS.md
@@ -5,8 +5,10 @@
 * Added the `selection_metric` argument to `ARIMA()`, which allows more control
   over the measure used to select the best model. By default this function will
   extract the information criteria specified by the `ic` argument.
-* Added `refit()` method for NNETAR (#287, @Tim-TU)
-* Added `refit()` method for MEAN (#289, @Tim-TU)
+* Added `refit()` method for NNETAR (#287, @Tim-TU).
+* Added `refit()` method for MEAN (#289, @Tim-TU).
+* Added `generate()` method for AR, the `forecast()` method now supports
+  bootstrap forecasting via this new method.
 
 ## Bug fixes